rss_pipeline/requirements.txt
psycopg2-binary
openai
boto3
trafilatura
feedparser
bs4